Desktop/New folder/multi-agent-fitness-website/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
google-generativeai==0.3.2
chromadb==0.4.18
sentence-transformers==2.2.2
numpy>=1.26.0
aiofiles==23.2.1
pymongo==4.6.1
motor==3.3.2
